refactor(tsql): Extract and compile generator
